// src/lib/schema.ts
export const AppointmentSchema = z.object({
    doctor_id: z.string().min(1, "Select physician"),
    patient_id: z.string().min(1, "Select patient"),
    type: z.string().min(1, "Select type of appointment"),
    appointment_date: z.string().min(1, "Select appointment date"),
    time: z.string().min(1, "Select appointment time"),
    note: z.string().optional()
})
// ...
export const VitalSignsSchema = z.object({
    patient_id: z.string(),
    medical_id: z.string(),
    body_temperature: z.coerce.number({
        message: "Enter recorded body temperature"
    }),
    heartRate: z.string({ message: "Enter recorded heartbeat rate" }),
    systolic: z.coerce.number({
        message: "Enter recorded systolic blood pressure"
    }),
    diastolic: z.coerce.number({
        message: "Enter recorded diastolic blood pressure"
    }),
    respiratory_rate: z.coerce.number().optional(),
    oxygen_saturation: z.coerce.number().optional(),
    weight: z.coerce.number({ message: "Enter recorded weight (Kg)" }),
    height: z.coerce.number({ message: "Enter recorded height (Cm)" })
})